2 Player keyboard movement and collisions
Players can move around using WASD and tab for player 1 and arrows and control for player 2. Collisions between ground, items and players are there (may want to see if we can get the players to be able to push other players). Cannot yet interact with items but planning to use area3D to see if items are close :)