Expanded Recipe UI
Addition of 3 more labels to display Recipes + Ingredients.
This uses a slot system where when a recipe becomes available they're assigned an ID and then that ID is assigned to one of the 4 Labels.
When one of the Recipe's timer hits 0 (can be changed to also include when it's turned in) it removes the recipe from the labels and makes them all move one slot back.